Add files via upload
This is a python code I wrote to increase the challenge on simple calculator. In this code I have tried to handle the errors which occur from either invalid input or incase the user is trying to divide a number by zero. Feel free to reach out to me if you find any mistakes that I can work on. Looking forward to responses.